requirements.txt
django
djangorestframework
rest_framework-simplejwt
mysqlclient
django-filter
